+++ app/models/subprogram.py
@@ -1,0 +1,20 @@
+from app.models.db import db
+
+
+class Subprogram(db.Model):
+    __tablename__ = 'sub_programs'
+    __table_args__ = {"schema": "pdl"}
+
+    id = db.Column(db.Integer, primary_key=True)
+    program_id = db.Column(db.Integer)
+    name = db.Column(db.String())
+
+    def __init__(self, program_id, name):
+        self.program_id = program_id
+        self.name = name
+
+    def __repr__(self):
+        return 'Subprograms(program_id=%s, name=%s)' % (self.program_id, self.name)
+
+    def json(self):
+        return {'program_id': self.program_id, 'name': self.name}