+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.13.0] - 2024-06-03
+
+### Changed
+- Update Title-1 endpoints to use database calls. [#182](https://github.com/policy-design-lab/pdl-api/issues/182) 
+- All programs and summary JSON files based on already mapped data. [#188](https://github.com/policy-design-lab/pdl-api/issues/179)
+
 ## [0.12.0] - 2024-04-09
 
 ### Removed
@@ -126,6 +132,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Open API specification file. [#50](https://github.com/policy-design-lab/pdl-api/issues/50)
 
+[0.13.0]: https://github.com/policy-design-lab/pdl-api/compare/0.12.0...0.13.0
+[0.12.0]: https://github.com/policy-design-lab/pdl-api/compare/0.11.0...0.12.0
+[0.11.0]: https://github.com/policy-design-lab/pdl-api/compare/0.10.0...0.11.0
+[0.10.0]: https://github.com/policy-design-lab/pdl-api/compare/0.9.0...0.10.0
+[0.9.0]: https://github.com/policy-design-lab/pdl-api/compare/0.8.0...0.9.0
+[0.8.0]: https://github.com/policy-design-lab/pdl-api/compare/0.7.0...0.8.0
+[0.7.0]: https://github.com/policy-design-lab/pdl-api/compare/0.6.1...0.7.0
+[0.6.1]: https://github.com/policy-design-lab/pdl-api/compare/0.6.0...0.6.1
 [0.6.0]: https://github.com/policy-design-lab/pdl-api/compare/0.5.0...0.6.0
 [0.5.0]: https://github.com/policy-design-lab/pdl-api/compare/0.4.0...0.5.0
 [0.4.0]: https://github.com/policy-design-lab/pdl-api/compare/0.3.0...0.4.0